:construction: mq manager

+++ pkg/req_resp.go
@@ -14,7 +14,7 @@ type TaskBaseData struct {
 	Database    string      `json:"database"`
 	Tables      []string    `json:"tables"` // default: all table
 	ShieldTable []string    `json:"shield_table"`
-	StartTime   int64       `json:"start_time"` // default: Starting from 0  , Use the latest: 1
+	StartTime   uint32      `json:"start_time"` // default: Starting from 0  , Use the latest: 1
 
 	ServerID     uint32 `json:"-"`
 	PositionName string `json:"-"`
